Move utils from disputes.ts into network.ts
